+++ Yvand.EntraCP.Tests/RequireExactMatchTests.cs
@@ -1,57 +1,34 @@
-﻿using Newtonsoft.Json;
-using NUnit.Framework;
+﻿using NUnit.Framework;
 using System;
-using System.Diagnostics;
 
 namespace Yvand.EntraClaimsProvider.Tests
 {
     [TestFixture]
-    public class RequireExactMatchOnBaseConfigTests : EntityTestsBase
+    [Parallelizable(ParallelScope.Children)]
+    public class RequireExactMatchOnBaseConfigTests : ClaimsProviderTestsBase
     {
-        public override void InitializeConfiguration(bool applyChanges)
+        public override void InitializeSettings()
         {
-            base.InitializeConfiguration(false);
+            base.InitializeSettings();
             Settings.FilterExactMatchOnly = true;
-            if (applyChanges)
-            {
-                GlobalConfiguration.ApplySettings(Settings, true);
-                Trace.TraceInformation($"{DateTime.Now.ToString("s")} [RequireExactMatchOnBaseConfigTests] Updated configuration: {JsonConvert.SerializeObject(Settings, Formatting.None)}");
-            }
+            Settings.ClaimTypes.UpdateIdentifierForGuestUsers(Configuration.DirectoryObjectProperty.UserPrincipalName);
+            base.ApplySettings();
         }
 
-        [Test, TestCaseSource(typeof(SearchEntityDataSource), "GetTestData", new object[] { EntityDataSourceType.AllAccounts })]
-        [Repeat(UnitTestsHelper.TestRepeatCount)]
-        public override void SearchEntities(SearchEntityData registrationData)
+        [Test]
+        public override void CheckSettingsTest()
         {
-            base.SearchEntities(registrationData);
+            base.CheckSettingsTest();
         }
 
-        [TestCase(@"aadgroup1143", 1, "3f4b724c-125d-47b4-b989-195b29417d6e")]
-        public override void SearchEntities(string inputValue, int expectedResultCount, string expectedEntityClaimValue)
+        [Test, TestCaseSource(typeof(EntraIdTestUsersSource), nameof(EntraIdTestUsersSource.GetTestData), null)]
+        public void TestAllEntraIDUsers(EntraIdTestUser user)
         {
-            base.SearchEntities(inputValue, expectedResultCount, expectedEntityClaimValue);
-        }
-    }
-
-    [TestFixture]
-    public class RequireExactMatchOnCustomConfigTests : CustomConfigTestsBase
-    {
-        public override void InitializeConfiguration(bool applyChanges)
-        {
-            base.InitializeConfiguration(false);
-            Settings.FilterExactMatchOnly = true;
-            if(applyChanges)
-            {
-                GlobalConfiguration.ApplySettings(Settings, true);
-                Trace.TraceInformation($"{DateTime.Now.ToString("s")} [RequireExactMatchOnCustomConfigTests] Updated configuration: {JsonConvert.SerializeObject(Settings, Formatting.None)}");
-            }
-        }
-
-        [Test, TestCaseSource(typeof(SearchEntityDataSource), "GetTestData", new object[] { EntityDataSourceType.AllAccounts })]
-        [Repeat(UnitTestsHelper.TestRepeatCount)]
-        public override void SearchEntities(SearchEntityData registrationData)
-        {
-            base.SearchEntities(registrationData);
+            // Input is not the full UPN value: it should not return any result
+            TestSearchOperation(user.UserPrincipalName.Substring(0, 5), 0, String.Empty);
+            // Input is exactly the UPN value: it should return 1 result
+            TestSearchOperation(user.UserPrincipalName, 1, user.UserPrincipalName);
+            TestValidationOperation(UserIdentifierClaimType, user.UserPrincipalName, true);
         }
     }
 }